demo app with:
selecting a region and you searching a summoner name in the home screen
clicking on the search button will trigger an action, which will update a reducer, saving the summoner name and region to it
TODO: get the route params somehow on the dashboard screen, and make the API call before we render it. This will also handle the case where the user enters in the URL directly and not through search (I will demonstrate later)

clicking the button will also redirect you to the route for dashboard, which will show your summoner name based on what's in the reducer. TODO: show it based on your route instead!

note: make sure you run npm install!